Change: DataAcquisitionView now allows the user to Select a Project
Before, Project was a requirement to show a DataAcquisitionView, but since we've now realised that it should be on its own Tab, we made it more independent.
This has meant that @appData needs to hold a list of projects again, and it's a bit redundant compared to the ViewState of ProjectsList. Maybe we can change that now.