Fixed NavBar for macOS
Before this change, on macOS we had two panes on the right side, and it was awkward. macOs doesn't want 'child views' to have their own NavigationView, but however, we need them for iOS. So we removed them from each Tab's view, and added a new wrapper helper in View that only adds a NavigationView under iOS.
So we have the best of both worlds.