AppData's updateResources() now downloads all Resources
We've been stuck for a long time trying to use a custom assign / sink Publisher/Subscriber, because the default one doesn't work, since it calls onError() because of the sink's completion getting called but without a failure.
We've gone the long way around, but at least there's something working here.