Big 'Views' Code Re-organisation
Since Dashboard is no longer our 'starting point', we needed to put everything back into a place where it made sense to find it. There was also a lot of clutter with all the 'Layout Views', so I made a new Group for 'RootView' and started putting things were I thought it made more sense.