Restored Working Devices Scanner
The issue is that the 'Publish' Combine Signal was not going through when the Scanner was buried inside DeviceData, so I reversed the effect, and since both DeviceData and Scanner work on Device(s), it made sense to unify everything into DeviceData, because past experience says this will work for us, and yeah. It's working as expected again.