Fix: Devices' Scanner now works on Startup, instead of only from the Second time onwards
This was something we were dragging since we started Scanning. The first call to a CBManager's State is the one that triggers the BLE Radio to turn on if present, but it does not turn it on immediately. So you can't just make two consecutive calls to it, because the second will fail - we need to wait a few cycles. What we're doing is that onAppear() of DeviceList, we call it through this new API that we called turnOnBluetoothRadio(), which just asks the CBCentralManager for its state without caring for the result. Then, when the user decides to start Scanning, the radio should be on, and if it's not we will get a valid error this time, because we did try to turn the radio on before.