Improved LoginState / AppData API
The @showingUser() enum for the logged-in state wasn't a good name, so now it's called "complete". Also, we had multiple places in the app where we checked for the User and Projects through switching on the loginState, so now they're all just if-checks through the new @user property, like we did for @projects.