Modified CollaboratorsDisclosureView
We wrote this code for the 'Project Selection' Dropdown, which is attempting to stack the Images horizontally. In the end, I think the code sort-of-works, and we want to use it, so we do but only for iOS, since on the Mac we just list the Collaborators. It's not its own Component because we're only using it one place. For now. Also, the math is not right regarding the Trailing padding.