Implemented 'About nRF Edge Impulse' Panel on macOS
We also fixed the name of 'Edge Impulse', which was previously 'nRF-Edge-Impulse' instead of 'nRF Edge Impulse'. Also, I noticed "CFBundleDisplayName", at least on macOS, does not return the expected value, but in the About Pane, it does show up well, which is SwiftUI / AppKit reading the value well. I don't know...