Changed some DataAcquisitionView-related Enums into Identifiables
This is because Roshan discovered a crash in iOS 14.1 if we switch the Sensor to 'Camera' and then back to 'Accelerometer'. The crash is mostly all UITableView gibberish and none of it is our code, but I'm going to make a couple of changes to see if we can address it. It seems to be 'Picker-related'.