Fix for DataAcquisitionView crash when switching Sensor to Camera and back
This crash for what we know happens in 14.1 and not in 14.4, and we tracked it down for the setAsComboStyle() helper. We've skipped this style for this particular version of iOS and we have the fix implemented in the helper function, so we don't have this crash happening again with us having to fix it ourselves.