Change: Project Sample Data is now held by AppData
We could not keep it locked in an @ObservableObject private only to DataSampleView, because it doesn't live long enough. So the solution was to put this in AppData, and this was quick to show results. We even managed to make it work when switching between Projects, since we trigger a request for new samples using a didSet {} listener.