Fixed 'New Sample' for macOS
The code we had using the toolbar to trigger a NavigationLink does not work on macOS, because the Toolbar does not fall underneath a NavigationView. We might be able to fix this, I don't know but, this workaround seems stable enough. It's a hack of sorts but, it gets us going forward.