Possible Fix for Logout not working
Logout was failing because the apiToken entry in the Keychain was not being deleted. It might be because we lacked permission, if another app uses the exact same String for some reason. So we added a prefix with the app's name, so only we (Edge Impulse) should have control of it.
I hope.