Fix: Renaming a Device doesn't make the new name show up in DevicesList
The Request is fine, and the data is fine. Problem is that if the ID is equal, then SwiftUI doesn't update it because there is no change to the underlying data. So for this change to be seen, we need to make Device(s) not equal if their names have been changed.