DeploymentView for iOS is very close
We got the form at the bottom, it's fixed Size, no scrolling via Introspect, we switch to show the Logs at the top, and... I can't get it to scroll. But I've fixed a lot of things and now I need a break from SwiftUI, because it's super annoying sometimes.