Fix for RenameDeviceView's TextField not becoming First Responder on iOS
This happened on iOS only. Under macOS it worked. The worst part was that I remember this working flawlessly, so it got me confused. However, thanks to the help from this issue posted on GitHub https://github.com/siteline/SwiftUI-Introspect/issues/6 I was able to reverse things. It's rather a blunt instrument fix, but it could make sense for other Dialog(s) / Alerts.