Fix: When connection to a RegisteredDevice fails, its State is now restored as .readyToConnect
The bug was that in this case at least, the handler's @device property is always nil, so we will never set it to the correct state. Instead, we must search for the registeredDevice belonging to that ScanResult, because we do have the @scanResult, and then we can restore the state to ready.