Fix: DeploymentView's DeploymentViewState is now a @StateObject instead of an @ObservedObject
Before this change, if the user changed Projects, the Socket and the Job status would be torn down. This is because as an ObservedObject, with the View 'redraw' it got destroyed. But it's state, so it needs to be maintained. thus the change. Now we can switch Projects and the state doesn't break.

Also, we need this to be @StateObject and not @State, because @State doesn't publish changes to its values, so our switch depending on the job's status enum doesn't trigger UI changes like @ObservedObject did.