Fix for DeploymentLogView crash in macOS
It crashed because of our Introspect trick didn't fly with AppKit, for memory / dead NSTableView reasons.
In the end, we gave our old friend ScrollViewReader another try, and maybe now we understand SwiftUI a bit better? It just worked once we figured out we were not getting the last item of the @logMessages array.