Fix for other crash in DeploymentLogView
The crash this time is because there probably was multiple copies of the same String. So instead of using String, we now made a new LogMessage struct that holds the log messages. And since it adds a timestamp, two messages with the same value should no longer share the same hash value.