We've got Keyboard Shortcuts for macOS too
We had to go through Commands, and also enable the selection binding for the Tab on the Navigation Link. I was not sure how dangerous that'd be, but so far, so good. It also fixed the problem of the app starting without having any of its Tabs selected. Good progress.