Fix for Deployment always failing on macOS
The issue is that everything is processed inside the UI Thread, and so if we keep scrolling the NSTableView on every new Message, we're stalling the thread from processing the WebSocket. If the WebSocket times out and cuts us off, we're going to fail, because we don't get all the data. So by debouncing the UI calls, we give more time for the WebSocket to read all the data.