Scan (gatherData) changes
Scan (reference: https://heckj.github.io/swiftui-notes/#reference-scan) is an Accumulator. It does not make much sense to try to decode part of it as a JSON and discard the other part, since JSONDecoder does not work with JSON fragments. It's the same if we just accumulate / append all Data as it is received. There's no change to our uses of gatherData() with this change.