We are now able to properly process the final 'body' Sample Data
The problem was that our gatherData() was adding (prepending) the previous JSON. So now, we filter it out before passing it out to gatherData to accumulate the next JSON, which contains the big body Data. As to why this happens, it's because the firmware fires the 'sampleUploading' and the final JSON with the full body in rapid-fire, so both are caught by our last gatherData() Accumulator.