Fix for DataAcquisition "Double-Pop-In" Issue
We created it with the IntrospectField. I think what happens is that we are interrupting the animation and propping up the keyboard. So by delaying the firstResponder, we don't interrupt the animation anymore.

By the way, we checked and the Mac is not affected. Also, the Mac has its own DataAcquisitionView.