Fix for DataAcquisitionView being cleared-out upon Successful Data Sampling
When we upload a Data Sample, we trigger a refresh of the list of Samples, so the user can see it. This triggers a refresh of the DataSamplesView, which re-creates the DataAcquisitionView. A simple changed fixed this, to my surprise.

Also: The set-to-false-then-true for the $isActive binding is not helping here. By moving the link outside the List, we fixed the issue with clicking the '+' and it not working. What a surprise. Mysteries of SwiftUI.