Fix for Duplicate Sampling Requests
Once we make a Sample Request ID, the WebSocket will kick us back in the ass to start Sampling. So if we send two requests, one ourselves, the other listening to the WebSocket, then we're in trouble. So yeah, we're going to RELY ON the WebSocket. Not dangerous nor risky at all.