Error Text Improvement
We were never able to get the text of the Error itself, because we needed to use LocalizedError and also switch the properties to implement. At least for our current use case that we were working on right now. But I'm confident this will improve the quality of the app everywhere else.