Small RegisteredDeviceRow Cleanup
I'm trying to fix the 'DeviceDetails' issue on macOS. I have an inkling / already something where it worked, but it didn't look right for other platforms. But we also need to keep things clean and not to mix changes between them so, first I had to get this out of the way.