Change: DeviceList is now a List, not a ScrollView, once again
The point of this change is to try to get NavigationLink(s) to work natively, instead of how we're forcing them now, because our current method is slow and makes the DetailsView take time to show up. This is not a fix, but I believe NavigationLink does not work as well as within a List - I could be wrong.

Also: In DeviceSection, we removed the Divider on the last item of the Section. I think this cleans up the UI a bit.