Fixed DeviceList Navigation
So, this is a big change. On iOS, it looks almost the same, as the previous code was attempting to make everything look like a Form. Now that we're using a Form (iOS-only), removing the custom drawing just leaves it looking quite good. This has not only made Navigation to DeviceDetails faster, it has also fixed the ContextMenu issue.

On the macOS side, it perhaps doesn't look as good as before, but indeed Navigation is quite fluid and fast now when opening DeviceDetails. So, one step at a time.