Fix for glitching Project Menu on iOS (disappears after a second or two when visible)
I wish something so simple and dumb didn't take so many hours to figure a workaround. Another fix was to declare the Button as a ToolbarItem from the TabBarView, but then we could not hide it for the other tabs. Sad.

I wonder now why we can set the '+' for Samples and not have any issues. SwiftUI, can you explain?