SSL (App Transport Security) Error Detection
We've been living with this bit of code for a while, and I couldn't fully test whether it was doing its job or not. The problem with ATS is that there's no 'immediate error', it just takes time to trip you out into an error, which is the real issue here.