Fixed Unregistered Device List Issues
We can now tap on any part of the Unregistered Device's cell, and it will trigger the 'tap' code. Also, the Progress Circle now shows up immediately, which was an issue of course of the underlying items now being 'different' (==) when the connecting state changes. That was also addressed here.