Fix for 'ConnectedDevicePicker' in macOS
We had to switch over the Device, not the Handler, for it to show up in macOS. This led to a cascade of fixes because we can't pull the remote handler from DeviceData from inside the ViewState, because it lives inside AppData, and so on and so on.
But I think we have it fixed here.