InferencingView now uses mono-spaced font for results as well as fixed-digit results
So, I learned this morning monospaced fonts are so all fonts share the same width, so in a grid of items, they all align. So, we switched over to using this for inferencing's grid of results, and set the min/max number of values to the same, so they align perfectly.