Fix for InferencingView not updating with default Scanner Settings (filtering not-connectable and non-UART devices)
This fix is different, because InferencingView is a root View, a Tab, so I can't inject it an environment object like we can do for Data Acquisition. So we went another route, which is to make AppData fire an 'object changed' message so the UI can listen to changes there. Thankfully, AppData's Cancellables are never cancelled or stop, or else this would stop working.