Enabled MultiPath TCP on iOS (not available for macOS)
MultiPath TCP allows iOS apps to fire requests on both Cellular and Wi-Fi, meaning if one of the network interfaces is wonky, the other one can pick up the slack without user interference. For example, it'd be awesome if the dumb folks at Instagram turned this on.

Anyway, Edge Impulse Studio folk apparently have no idea about this, and their suggestion was "turn it on and see what happens", so that's what we're doing.