Fixed Inferencing
It was not working, I think because JSON parsing failed to find the anomaly field. So we set it to being optional, because it might show up. But it was not a big deal. I also saw that on Android the space between the columns is much bigger, so I decided to separate the columns a lot more.