Sweep View ID changes in modifier functions & utilities
So, SwiftUI doesn't like branching to return different copies of the same View. Instead, it likes it if the same View is returned with different transformations applied to it depending on the situation. I think we were aware of this before, because I could not find many of these or any in the code. Still, we made some changes.