Changes to try to fix Arjan's UI issue on first login
Arjan's bug report is that when first logging-in to the app, the background, I suppose of Devices Tab, is see-through. So we've added more background 'layers' for both Tab and Dual-Pane. I tried to remove the backgrounds from the Tab views themselves so as to not be repeating code anymore, but for example in Device Acquisition we lost the background of the Picker. So I decided to leave everything as-is.