Improvement: 'Reset' or 'Applying' DeploymentStage now has a 45-second Timer
This matches the change Roshan made on the Android side, we 'wait' for the Image Swap to complete before marking the update as complete. I thought this would require more changes, if I'm honest. I was thinking of re-using the 'Speed' Indicator for the Time Remaining part, but I think I might leave it just as it is.