Improved UI Design for Account Deletion, as well as macOS
I thought this UI was split between iOS and macOS, but it is not. Anyway, macOS needed its own 'Account Section' design to look better, though it still looks pretty poor in my opinion, but anyway, it's sharing all the code and logic, so we can assume Account Deletion work. Also made a couple of improvements to the UI on the iOS side.