Deleted HeroView and more changes to macOS' UserContentView
Hero was not used anymore - it was a good idea to borrow from Tweetbot but we never got it to look quite right. So we pulled it and now it's a bit better with the 'rounded box' around it I think. Or perhaps I'm going too far with that, I can't tell properly.