Crash Fix when deleting a device?
I think the source of the crash is that the previous code did not hold a strong reference nor a weak one. So now we keep it all weak - anyway if the app or UI is gone the user won't care about the UI updating the state. Also, we're now showing the User why the request fails, if the Server returns such information.