Fix for Socket Job Result Regular Expression
The Regular Expression did not fail, but it did not isolate the 'false or true' string from it, because more text was added after it in subsequent back end / server changes. So the string comparison later down in the code failed because the extracted string had more stuff in it.