Deleted DeploymentStage and updated DeploymentProgressManager to use PipelineView
Of course, the existing APIs also had to be updated to support progress and the extra accessory line, which took some time. But the new API is now roughly up to par of where it needed to be.
Tested that macOS is also showing the Pipeline and progress as expected.