Fixed an Xcode warning in CollaboratorsHStackView
There isn't a quick-and-dirty way to fix this. We want to offset the items, so we need an index. We could use a zip of the indices, because apparently that's the way to do it instead of using enumerated, but I'd rather just keep it this way because the code makes it clear why we need the index.