Deleted / Removed BetterSafariView dependency
Apparently there's something called ASWebAuthenticationSession we can use. Only discovered it because App Store Review blocked or rejected the Mac version and asked us if we were using it as one of the reasons, the other was pretty minor. So I think this was a case where App Review was actually helpful. Hopefully.